MapService/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
websocket-client==1.5.1
Werkzeug==2.3.4
yarl==1.9.2
